requirements.txt
# Core ML and NLP
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.7.0
sentencepiece>=0.1.99
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
faiss-cpu>=1.7.4

# App and document processing
streamlit>=1.33.0
pypdf>=4.2.0

# Existing project dependencies
spacy>=3.6.0
bert-score>=0.3.13
rouge-score>=0.1.2
pandas>=2.0.0
datasets>=2.14.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1
joblib>=1.3.0
regex>=2023.6.3
accelerate>=0.21.0
